GoogleScholorSummary/requirements.txt
google-search-results
google-generativeai
requests
python-dotenv

